Removed dotenv and lambda-api dependencies
